refactor: migrate SourceType from useMetrics to useAnalytics
